requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiofiles==24.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
databases[postgresql]==0.9.0
asyncpg==0.30.0
email-validator==2.2.0 